Update README.md with correct `.graphqlconfig`
Hi there! Firstly, thank you for this project. As I was getting this set up for a project of my own, I noticed that there was a slight error in the example `.graphqlconfig` in the README. The `includes` field should be an array of globs, not just a string. The downstream `graphql-config` package basically does a call that looks like `(this.config.includes || []).map(...)`, and this obviously fails in the case of just a single string `includes` field. 

Let me know if I need to do anything else to get this small change through. Thanks!